refactor: separated the methods in two new crates for better code organization
